Add feature serde: (De)serialize datatypes

+++ src/commands/select.rs
@@ -13,7 +13,7 @@ pub fn parse(args: Vec<String>) -> Result<Request, RedisProtocolError>{
     }
     
     if let Some(arg) = args.get(0) {
-        if let Ok(number) = arg.parse::<usize>() {
+        if let Ok(number) = arg.parse::<u64>() {
             return Ok(Request::SELECT(number))
         }
     }
